moved jumppathing to partial class
